Cutscene, model fixes, path puzzle, bug fixes and secret surprise
Created a simple cutscene for the cave.
Fixed a lot of models'  pivots to work properly with the player, had to reimport the new model with the proper pivots and scaled some of the models up.
Fixed some  more bugs.
Made the path puzzle.
Also added one of the most essential things in a video game, its subtle but very much required in any video game.